feat(ivy): support separate .js and .d.ts trees when generating imports
The NgModule handler generates R3References for its declarations, imports,
exports, and bootstrap components, based on the relative import path
between the module and the classes it's referring to. This works fine for
compilation of a .ts Program inside ngtsc, but in ngcc the import needed
in the .d.ts file may be very different than the import needed between .js
files (for example, if the .js files are flattened and the .d.ts is not).

This commit introduces a new API in the ReflectionHost for extracting the
.d.ts version of a declaration, and makes use of it in the NgModule
handler to write a correct expression for the NgModule definition type.