refactor(ivy): split 'type' into 'type' and 'internalType' for defs
When compiling an Angular decorator (e.g. Directive), @angular/compiler
generates both an 'expression' to be added as a static definition field
on the class, and a 'type' which will be added for that field to the .d.ts
file.

Previously, the same WrappedNodeExpr of the class' ts.Identifier was used
within both 'expression' and 'type'.

In the ngtsc case, this is proper. In the ngcc case, if the class being
compiled is within an ES5 IIFE, the outer name of the class may have
changed. Thus, the class has both an inner and outer name. The outer name
should continue to be used elsewhere in the compiler and in 'type', but
since 'expression' will live within the IIFE, the inner name should be used.

This commit introduces `ReflectionHost.getInternalNameOfClass`, which
the compiler can use to query for the correct name to use internally within
'expression'. This name is then passed via a new `internalType` property on
all metadata.