feat: single-file compilation spike
This commit allows the Angular compiler to run in a hypothetical single-file
compilation mode. In this mode, errors are not reported and the compiler does
not attempt to follow the import structure for NgModules or standalone imports.
Instead, code is emitted to link together components and their template contexts
(e.g. NgModules) at runtime.

All of this is untested and only really serves as a proof-of-concept.