refactor(core): large restructuring of DI code
Despite the scary size, this commit largely just moves DI code around to
make it a bit more organized and accessible. It does make a few behavioral
changes that should be no-ops:

One, it standardizes on one concept of "current injector". This eliminates
the "inject implementation" global variable and the `directiveInject()`
instruction in favor of a `DIRECTIVE_INJECTOR` implementation of `Injector`.
The runtime code path is the same.

Two, it eliminates the legacy "limp mode" of using DI without a current
injector (which only ever supported `providedIn: 'root'` tokens). This mode
is unused in modern Angular and existed to support experiments around the
smallest possible Hello World application. Modern Angular can assume DI is
used.

Three, it renames `R3Injector` to `InjectorImpl`, and restructures its
inheritance hierarchy since there's no longer the complication of
`StaticInjector`.