perf(forms): replace naive `deepSignal` with optimized `nestedSignal`
In signal forms, the form model is stored in a root `WritableSignal`, which
drives a hierarchy of fields that derive their values from that model.
Updating any one field causes a cascading update to the root signal.

Previously, the resulting write to the root marks as dirty the entire signal
tree below it, even if most of the tree did not change. While the semantics
and fine-grained nature of the signal graph prevents recomputation and side-
effects from running still, the bookkeeping cost of processing these changes
can be meaningful for a large graph.

`nestedSignal` uses a more optimized algorithm that leverages the knowledge
of _what_ is changing in the hierarchical structure to not notify downstream
signals which don't need to recompute their values.