perf(forms): use `structuralSignal` for `childrenMap` derivation
`childrenMap` derives the structure of child `Field`s for a given node.
It is defined as a `linkedSignal` over the value of the node, which updates
a `Map` of children by diffing the incoming value.

Since `childrenMap` is only concerned with the presence/absence of child
fields and not their values, it's a perfect use case for `structuralSignal`.
This optimization allows `childrenMap` to avoid recalculations which
wouldn't affect its result, such as the updating of values within children.
These updates still change the value signal of the current field, but
`structuralSignal` avoids signal graph notifications when the structure of
the current value remains the same.