fix(forms): remove `PathKind`, `keyInParent` -> `key`, add `index`
`PathKind` added a lot of type complexity, and still was broken as it only
narrowed `FieldContext` and not `FieldState`. Even more type manipulation would
have been required to make `FieldState` narrow correctly.

This commit removes `PathKind` in favor of a simplified `key()` and `index()`
pair of signals on `FieldState`, both of which throw if called against the
wrong kind of parent.