refactor(forms): replace `SchemaPath`'s `TSupportsRules` with new type
Previously, `SchemaPath<T>` referred to a `SchemaPath` that required rules
support, and `SchemaPath<T, SchemaPathRules>` referred to a path that did
not require rules support (because `SchemaPathRules` could be either
`Supported` or `Unsupported`). This works, but is confusing and creates type
complexity for declared validators.

This commit aims to simplify schema pathing by making `SchemaPath`s _always_
require rule support, and introducing `SchemaReferencePath<T>` to request a
path that does not require rule support. Thus, every `SchemaPath` is a
`SchemaReferencePath`, but not the other way around.

For example, `FieldContext.stateOf` accepts a `SchemaReferencePath` because it
can function for any path you can reference, not just paths which support
adding new rules.